+++ CHANGELOG.md
@@ -3,11 +3,37 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+No changes
+
+## [2.3] - 2018-03-14
+### Changed
+- Fix multi-window support
+
+## [2.2] - 2018-03-14
+### Changed
+- Fix Popup tab URL lookup
+
+## [2.1] - 2018-03-14
+### Changed
+- Fix Regular link in popup
+- Fix Chromium bug around adding event listener to null delete button
+
+## [2.0] - 2018-03-13
+### Changed
+- Switch to just blacklist of sites to run on
+
+### Added
+- Browser action popup to add rule
+- Background script to determine activeness
 
 ## [1.1] - 2016-08-29
 ### Added
 - This CHANGELOG file
 - Copy event blocking prevention
 
-[Unreleased]: https://github.com/jswanner/DontFuckWithPaste/compare/v1.1...HEAD
+[Unreleased]: https://github.com/jswanner/DontFuckWithPaste/compare/v2.3...HEAD
+[2.3]: https://github.com/jswanner/DontFuckWithPaste/compare/v2.2...v2.3
+[2.2]: https://github.com/jswanner/DontFuckWithPaste/compare/v2.1...v2.2
+[2.1]: https://github.com/jswanner/DontFuckWithPaste/compare/v2.0...v2.1
+[2.0]: https://github.com/jswanner/DontFuckWithPaste/compare/v1.1...v2.0
 [1.1]: https://github.com/jswanner/DontFuckWithPaste/compare/v1.0...v1.1